jwt token authentication for user login details
