--- middlewares/authMiddleware.js
@@ -1,5 +1,0 @@
-import jwt from "jsonwebtoken";
-
-export const authenticateToken = (req, res, next) => {
-	const token = req.headers["authorization"];
-};
